Enable syncing GCM readings with Services
For someone this is set to disabled, which causes GCM readings not to be uploaded to services like Tidepool or Nightscout.

If this is because one may be uploading from Libre to these services already directly, this should be turned into a setting that can be toggled from the settings view.